requirements.txt
blinker=1.9.0
click=8.1.8
colorama=0.4.6
Flask=3.1.2
Flask-MQTT=1.2.1
importlib_metadata=8.7.0
itsdangerous=2.2.0
Jinja2=3.1.6
MarkupSafe=3.0.3
paho-mqtt=1.6.1
pip=25.3
setuptools=57.4.0
Werkzeug=3.1.3
zipp=3.23.0